### 2025-02-05 #7652 @mbkumar
Yes.

Bharat Medasani

On Tue, Feb 4, 2025 at 8:31 PM Rachel Kurchin ***@***.***>
wrote:

> 👋 @jranalli <https://github.com/jranalli>, @mbkumar
> <https://github.com/mbkumar>, and/or @loftytopping
> <https://github.com/loftytopping>, would you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
