### 2025-02-05 #7751 @editorialbot
[REVIEW]: 'KinOpt: A Python package for chemical kinetics analysis and optimization'
**Submitting author:** @alan-tabore (<a href="http://orcid.org/0009-0009-6666-7207">Alan Taboré</a>)
**Repository:** https://github.com/alan-tabore/KinOpt
**Branch with paper.md** (empty if default branch): joss_branch
**Version:** v1.0
