### 2025-02-07 alan-tabore/KinOpt#1 @jranalli
Hiccups with installation instructions
First off, I will say that I'm a US-based Windows user and not a multiplatform expert, so if there's some of this stuff that is platform or OS version dependent, I apologize in advance.

### Pip command issues
The installation instructions given for `pip` don't work.
`python -m pip install requirements.txt`
I think you want:
`python -m pip install -r requirements.txt`

### Potential python version issues
I don't have familiarity with all the dependencies you're using, so I'm not sure where the issue cropped up. But my initial attempt to install with pip using python 3.12 on Windows failed with an error about requiring specific python versions for one package. Tried again with python 3.10 and installation worked with pip. See more details below.

### Conda command issues
Similarly to pip, I believe you need a different command here
`conda install --file requirements.txt`

### File encoding issues for Conda
After downloading the code, the requirements.txt file had a UTF-16LE encoding. That didn't work out of the box on my US Windows 11 PC (again, sorry if there are platform/environment dependent issues on this, I'm not an expert at that stuff). pip handled it but conda had trouble and threw errors that involved mangled package names. Manually changing the encoding to UTF-8 in my IDE allowed it to be read correctly, but...

### Conda packages not found. Possible channel/version issue?
Still couldn't get the conda install command to work even after fixing the encoding. Conda couldn't find all the packages (I tried both with default channels and adding in conda-forge). Did you use some extra channels that you'd need to specify?
```
PackagesNotFoundError: The following packages are not available from current channels:

  - wcwidth==0.2.13
  - traitlets==5.14.1
  - shiboken6==6.6.2
  - shiboken2==5.15.2.1
  - pyside6_essentials==6.6.2
  - pyside6_addons==6.6.2
  - pyside6==6.6.2
  - pyside2==5.15.2.1
  - pyqt5designer==5.14.1
  - pyqt5-qt5==5.15.2
  - pyqt5==5.15.10
  - pyparsing==3.1.1
  - plotly==5.20.0
  - matplotlib==3.8.3
  - kiwisolver==1.4.5
  - fonttools==4.49.0
  - et-xmlfile==1.1.0
  - cycler==0.12.1
  - tzdata==2024.1
```

In some of these cases the package is there, e.g. `conda install plotly` works, so I think it might be an issue with the specific versions you're requesting rather than the packages themselves?
