### 2025-02-07 alan-tabore/KinOpt#2 @jranalli
Tests failing
I had three of the tests fail for me. They appear to all be based on hard-coded data, so I don't think it's an issue on my end.

``` python
KinOpt\tests\test_kinetic_models.py:10 (test_vitrification_WLF_rate_no_reaction_below_Tg)
def test_vitrification_WLF_rate_no_reaction_below_Tg():
        # Test case 1: All temperatures above Tg
        temperature = np.array([350, 400, 450])
        Ad = 1.0
        C1 = 0.5
        C2 = 0.2
        Tg = 320
        expected_result = np.array([1.64327096, 1.64666679, 1.64745546])
        result = km.vitrification_WLF_rate_no_reaction_below_Tg(temperature, Ad, C1, C2, Tg)
>       assert np.allclose(result, expected_result), "Test case 1 failed"
E       AssertionError: Test case 1 failed
E       assert False
E        +  where False = <function allclose at 0x000001DD5C006370>(array([0.55498602, 0.55869046, 0.56193328]), array([1.64327096, 1.64666679, 1.64745546]))
E        +    where <function allclose at 0x000001DD5C006370> = np.allclose
```

```
KinOpt\tests\test_optimization.py:32 (test_rss_increase_of_small_rates_impact_with_zones)
def test_rss_increase_of_small_rates_impact_with_zones():
        rss = opt.rss_increase_of_small_rates_impact_with_zones(x, experimental_rate, mock_rate_law, (), 0, mock_vitrification_law, (), 0, mock_coupling_law, (), mock_tg_law, (), (), fraction_to_amplify, amplification_factor)
>       assert np.isclose(rss, 0.025, atol=1e-3)
E       assert False
E        +  where False = <function isclose at 0x000001DD5C0064F0>(1.6099999999999999, 0.025, atol=0.001)
E        +    where <function isclose at 0x000001DD5C0064F0> = np.isclose
```

```
KinOpt\tests\test_optimization.py:37 (test_rss_increase_of_small_extents_impact)
def test_rss_increase_of_small_extents_impact():
        extent = np.array([0.05, 0.1, 0.15, 0.2])
        extent_limit = 0.1
        rss = opt.rss_increase_of_small_extents_impact(x, experimental_rate, mock_rate_law, (), 0, mock_vitrification_law, (), 0, mock_coupling_law, (), mock_tg_law, (), (), extent, extent_limit, amplification_factor)
>       assert np.isclose(rss, 0.0325, atol=1e-3)
E       assert False
E        +  where False = <function isclose at 0x000001DD5C0064F0>(0.47, 0.0325, atol=0.001)
E        +    where <function isclose at 0x000001DD5C0064F0> = np.isclose
```
