### 2025-02-07 alan-tabore/KinOpt#3 @jranalli
Missing Community Guidelines
JOSS requires Community Guidelines to be listed. See description [here](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines).

These need not be extensive, you could consider what's listed on other small projects, e.g. [solarspatialtools](https://github.com/jranalli/solarspatialtools).
