### 2025-02-07 alan-tabore/KinOpt#4 @jranalli
Automated Tests Lack Instructions
The project has tests (though they are failing, see Issue #2 ). But JOSS requests that you describe [how to run them](https://joss.readthedocs.io/en/latest/review_criteria.html#tests).

You may wish to consider implementing Continuous Integration using Github Actions to run these tests automatically with each commit, but that's not required for JOSS.
