### 2025-02-07 alan-tabore/KinOpt#5 @jranalli
Suggestion: Consider default options
I admittedly don't know the science of the kinetics well enough to understand the interplay between the model and the optimizer choices. It might be useful to fill some default options or make recommendations of which to start with for non-experts, because I did notice drastically different performance (i.e. 1hr+ running time) with one combination that I chose at random (once with the basinhopping and once with the shgo optimizer).

Given that, one additional suggestion might be to have a cancel button in the GUI for the optimization to stop hour-plus runs.
