### 2025-02-07 #7751 @jranalli
I ran into a few issues, so here's the concerns I have.

### Must Fix - areas that failed to meet JOSS criteria:
- [ ] Some issues with installation instructions, detailed in [Issue Link](https://github.com/alan-tabore/KinOpt/issues/1)
- [ ] Community Guidelines: No community guidelines or contribution instructions. This doesn't need to be extensive, just seek examples from small projects [Issue Link](https://github.com/alan-tabore/KinOpt/issues/3).
- [ ] Functionality/Automated Tests: three of the included tests failed detailed in [Issue Link](https://github.com/alan-tabore/KinOpt/issues/2).
- [ ] Automated tests: the project doesn't use CI and lacks explicit instructions for running the tests. [Issue Link](https://github.com/alan-tabore/KinOpt/issues/4).

### My opinions that aren't necessary to fix to meet JOSS criteria but may be viewed as constructive suggestions
- [ ] Include some default options. [Issue Link](https://github.com/alan-tabore/KinOpt/issues/5)
- [ ] Paper: The paper is well written. I think there could be some improvement with respect to its accessibility for a "non-sepcialist audience."
- [ ] Example usage: the API docs based on the dosctrings are good, and the video tutorials for the GUI are good, but it would be nice to include an example of how to perform the analysis from code/script.
- [ ] Consider improving the code coverage to include some of the other modules.
