### 2025-02-08 alan-tabore/KinOpt#1 @alan-tabore
Hello,
No need to apologize, this software is intended for beginner users in programming (so mostly using Windows), so it should be trivial for them to install the package software.

As you can see, I'm no expert in packaging but I'll try to create a pip package, i.e:
` pip install kinopt `
to facilitate the installation of the package.
If possible, I'll also automate the verification of dependencies versions with GitHub Actions
