### 2025-02-08 alan-tabore/KinOpt#2 @alan-tabore
Thanks for your feedback, I'll improve these tests using more trivial/appropriate hard values (and also explain the goal of those tests).
