### 2025-03-05 alan-tabore/KinOpt#1 @mbkumar
I'll give my observations regarding installation in this thread instead of creating a duplication issue.

I tried both Mac and Linux and I couldn't install the software with either pip or conda.  As pointed by @jranalli, the installation instructions need to be corrected. Also the target OS should be specified in the docs.
