### 2025-05-22 c4d728c
Create CONTRIBUTING.md
