### 2025-05-22 #7751 @alan-tabore
Hello,

I just added a community guidelines.
I'll try to create a pip package before the end of the week.
