### 2025-05-28 8091874
Removing shiboken2 dependencies from requirements.txt
