### 2025-05-28 alan-tabore/KinOpt#1 @alan-tabore
I corrected the installation instructions and requirements.
The installation process should work now.

I also tried installing everything with python 3.11 and except for the warning after running:
``` bash
python -m pip install -r requirements.txt
```
there was no trouble running the installation using pip.

I don't want to specify a specific OS as I want this project to be used on Windows/Mac/Linux.
