### 2025-06-09 #7751 @alan-tabore
Hi, I modified the instructions of installation and everything should be working now.
I'm currently modifying the tests.
