### 2025-06-11 8ccffad
correction of tests for vitrification functions
