### 2025-06-11 487a821
adding test function to verify input arguments of function defined in the kinetic_models module
