+++ kinopt/tests/test_kinetic_models.py
@@ -6,7 +6,8 @@ Created on Thu Jun 15 11:39:41 2023
 """
 import pytest
 import numpy as np
-import src.kinetic_models as km
+import inspect
+from kinopt.src import kinetic_models as km
     
 def test_vitrification_WLF_rate_no_reaction_below_Tg():
     # Test case 1: All temperatures above Tg
@@ -40,7 +41,34 @@ def test_vitrification_WLF_rate_no_reaction_below_Tg():
     result = km.vitrification_WLF_rate_no_reaction_below_Tg(temperature, Tg, Ad, C1, C2)
     assert np.allclose(result, expected_result), "Test case 4 failed"
 
-    print("All test cases passed!")
+
+def test_rate_functions_signature():
+    for name, func in inspect.getmembers(km, inspect.isfunction):
+        if name.startswith("rate_"):
+            sig = inspect.signature(func)
+            params = list(sig.parameters.values())
+            assert len(params) >= 2, f"{name} should have at least 2 parameters"
+            assert params[0].name == "extent", f"{name}: first argument should be 'extent'"
+            assert params[1].name == "T", f"{name}: second argument should be 'T'"
+
+def test_vitrification_functions_signature():
+    for name, func in inspect.getmembers(km, inspect.isfunction):
+        if name.startswith("vitrification_"):
+            sig = inspect.signature(func)
+            params = list(sig.parameters.values())
+            assert len(params) >= 2, f"{name} should have at least 2 parameters"
+            assert params[0].name == "T", f"{name}: first argument should be 'T'"
+            assert params[1].name == "Tg", f"{name}: second argument should be 'Tg'"
+
+def test_tg_functions_signature():
+    for name, func in inspect.getmembers(km, inspect.isfunction):
+        if name.startswith("tg_"):
+            sig = inspect.signature(func)
+            params = list(sig.parameters.values())
+            assert len(params) >= 1, f"{name} should have at least 1 parameter"
+            assert params[0].name == "extent", f"{name}: first argument should be 'extent'"
+
+print("All test cases passed!")
     
     
 if __name__=="__main__":
