### 2025-06-11 c537187
remove test functions of the optimization module
