### 2025-06-17 f835c18
Create python-app.yml
