### 2025-06-17 f8b8bd4
Update requirements.txt to have a working version of PyQt Designer for ubuntu users (pyqt5-tools)
