+++ requirements.txt
@@ -17,7 +17,7 @@ pyparsing==3.1.1
 PyQt5==5.15.10
 PyQt5-Qt5==5.15.2
 PyQt5-sip==12.13.0
-pyqt5-tools==5.15.4.3.2
+pyqt5-tools==5.15
 PySide6==6.6.2
 PySide6_Addons==6.6.2
 PySide6_Essentials==6.6.2
