### 2025-06-17 alan-tabore/KinOpt#1 @alan-tabore
Everything should be working, I also added a Github Action to check that the build of the package is correct
