### 2025-06-17 alan-tabore/KinOpt#3 @alan-tabore
Community Guidelines have been added in the form of a _contributing.md_ file.
