### 2025-06-17 alan-tabore/KinOpt#5 @alan-tabore
Thanks for your recommendations.

For the optimization method, I provided some "auto-fill values" but regarding the optimization algorithm some recommendations could prove useful. My advice would be to use a basin-hopping algorithm with a local optimization performed with the Nelder-Mead algorithm (using default values for both). This optimization can take up to a few hours but ensures that you're not stuck in a local minimum.
To be fair, the interplay between the model and optimizer is not well documented. That's why I wanted the user to be able to play with the various models and optimization methods.

The cancel button is also a must have. I'll let this issue open until the cancel button is added.
