### 2025-06-18 7f3d3bb
Adding test for the interpolation module
