### 2025-06-18 cba6394
removing useless test
