### 2025-06-20 60e4fd4
Removing unused function
