+++ kinopt/src/kinetic_models.py
@@ -661,93 +661,6 @@ def compute_extent_and_rate(time, temperature, rate_law=None, rate_law_args=None
 
 
 
-
-def compute_extent_and_rate_using_scipy(t0, tf, temperature_program, rate_law, rate_law_args, vitrification_law=None, vitrification_args=None, tg_law=None, tg_law_args=None, coupling_law=None, coupling_law_args=None, initial_extent=0):
-    """
-    Compute extent and rate using scipy's solve_ivp.
-
-    Parameters
-    ----------
-    t0 : float
-        Initial time.
-    tf : float
-        Final time.
-    temperature_program : callable
-        A function that takes time as input and returns temperature.
-    rate_law : callable
-        Rate law function.
-    rate_law_args : tuple
-        Arguments for the rate law function.
-    vitrification_law : callable
-        Vitrification law function.
-    vitrification_args : tuple
-        Arguments for the vitrification law function.
-    tg_law : callable
-        Tg law function.
-    tg_law_args : tuple
-        Arguments for the Tg law function.
-    coupling_law : callable
-        Coupling law function.
-    coupling_law_args : tuple
-        Arguments for the coupling law function.
-    initial_extent : float, optional
-        Initial extent. Default is 0.
-
-    Returns
-    -------
-    scipy.integrate.OdeResult
-        Solution object from solve_ivp.
-    """
-    if (vitrification_law is not None) and (tg_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-    if (vitrification_law is not None) and (coupling_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-    if (coupling_law is not None) and (vitrification_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-    if (coupling_law is not None) and (tg_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-    if (tg_law is not None) and (vitrification_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-    if (tg_law is not None) and (coupling_law is None):
-        raise ValueError("Please make sure to indicate a tg law since you have selected a vitrification law.")
-
-    def rate(t, extent, temperature, *args):
-        """
-        Compute the rate at a given time.
-
-        Parameters
-        ----------
-        t : float
-            Time.
-        extent : float
-            Current extent.
-        temperature : float
-            Current temperature.
-        *args : tuple
-            Additional arguments.
-
-        Returns
-        -------
-        float
-            Rate.
-        """
-        rate = rate_law(extent, temperature_program(t), *rate_law_args)
-
-        if vitrification_law is not None:
-            tg = tg_law(extent, *tg_law_args)
-            vitrification_term = vitrification_law(temperature_program(t), tg, *vitrification_args)
-            rate = coupling_law(rate, vitrification_term, *coupling_law_args)
-        
-        return rate
-    
-    initial_extent = np.atleast_1d(initial_extent)
-    return scipy.integrate.solve_ivp(rate, [t0, tf], initial_extent,
-                                     args=(temperature_program, *rate_law_args,
-                                           vitrification_law, *vitrification_args,
-                                           tg_law, *tg_law_args,
-                                           coupling_law, *coupling_law_args))
-
-
 #%% Example 1 - Kamal
 if __name__=="__main__":
     
