+++ kinopt/src/kinetic_models.py
@@ -21,11 +21,11 @@ def arrhenius_rate_constant(T,A,Ea):
     Parameters
     ----------
     T : Float
-        Temperature of reaction.
+        Temperature of reaction in Kelvin.
     A : Float
         Pre-exponential factor.
     Ea : Float
-        Activation energy.
+        Activation energy in J/mol.
 
     Returns
     -------
@@ -53,11 +53,11 @@ def rate_for_nth_order(extent,T,A1,E1,n):
     extent : 1-D array
         Extent of reaction.
     T : 1-D array
-        Temperature of reaction.
+        Temperature of reaction in Kelvin.
     A : Float
         Pre-exponential factor of the reaction.
     Ea : Float
-        Activation energy of the reaction.
+        Activation energy of the reaction in J/mol.
     n : Float
         Order of reaction.
 
@@ -67,8 +67,47 @@ def rate_for_nth_order(extent,T,A1,E1,n):
         Rate of reaction for a nth order reaction.
 
     """
-    return arrhenius_rate_constant(T,A1,E1)*(1-extent)**n
+    return arrhenius_rate_constant(T,A1,E1)*((1-extent)**n)
 
+
+def rate_for_autocatalytic(extent,T,A,Ea,m,n):
+    r"""
+    Compute the rate of reaction for an autocatalytic reaction.
+    
+    Parameters
+    ----------
+    extent : ndarray
+        Extent of reaction.
+    T : ndarray
+        Temperature of reaction in Kelvin.
+    A : float
+        Pre-exponential factor of the reaction.
+    Ea : float
+        Activation energy of the reaction in J/mol.
+    m : float
+        Order of reaction for the autocatalyzed reaction.
+    n : float
+        Order of reaction for the regular reaction.
+    
+    Returns
+    -------
+    rate : ndarray
+        Rate of reaction for an autocatalytic equation.     
+        
+    Notes
+    -----
+     The rate for an autocatalytic reaction is given by:
+    
+     .. math:: \frac{d\alpha}{dt} =  A e^{ \left( \frac{-E_a}{RT} \right)} \alpha^m (1-\alpha)^n
+     
+    References
+    ----------
+    [1] M. R. Keenan, « Autocatalytic cure kinetics from DSC measurements: Zero initial cure rate », J. Appl. Polym. Sci., vol. 33, nᵒ 5, p. 1725‑1734, avr. 1987, doi: 10.1002/app.1987.070330525.
+
+    """
+    if np.any(extent) <= 0:
+        raise ValueError("Be careful, for an autocatalytic model the extent can't be inferior or equal to 0 !!! \n It would result in a rate of reaction equal to 0 and no reaction would occur.")
+    return arrhenius_rate_constant(T,A,Ea) * extent**m * (1 - extent)**n
 
 
 def rate_for_kamal(extent,T,A1,E1,A2,E2,m,n):
@@ -80,15 +119,15 @@ def rate_for_kamal(extent,T,A1,E1,A2,E2,m,n):
      extent : ndarray
          Extent of reaction.
      T : ndarray
-         Temperature of reaction.
+         Temperature of reaction in Kelvin.
      A1 : float
          Pre-exponential factor of the regular reaction.
      E1 : float
-         Activation energy of the regular reaction.
+         Activation energy of the regular reaction in J/mol.
      A2 : float
          Pre-exponential factor of the autocatalyzed reaction.
      E2 : float
-         Activation energy of the autocatalyzed reaction.
+         Activation energy of the autocatalyzed reaction in J/mol.
      m : float
          Order of reaction for the autocatalyzed reaction.
      n : float
@@ -153,45 +192,6 @@ def rate_for_kamal(extent,T,A1,E1,A2,E2,m,n):
     """
[…]
