### 2025-06-20 7d5f41d
Adding multiple test functions for the kinetic_models module
