+++ kinopt/tests/test_kinetic_models.py
@@ -8,6 +8,81 @@ import pytest
 import numpy as np
 import inspect
 from kinopt.src import kinetic_models as km
+
+def test_arrhenius_rate_constant():
+    A = 1.0  # Pre-exponential factor
+    Ea = 831.446261815324  # Activation energy in J/mol
+    T = 100.0  # Temperature in Kelvin
+    
+    expected_result = np.exp(-1)
+    result = km.arrhenius_rate_constant(T, A, Ea)
+    assert np.allclose(result, expected_result), "Arrhenius rate constant calculation failed."
+   
+def test_rate_for_nth_order():
+    extent = np.array([0, 0.1, 0.3, 0.5, 0.6, 1])
+    T = np.array([100.0, 200.0, 300.0, 400.0, 500.0, 600.0])
+    A = 1.0  # Pre-exponential factor
+    Ea = 831.446261815324  # Activation energy in J/mol
+    n = 2  # Order of the reaction
+    
+    expected_results = [np.exp(-1), ((0.9)**2)*np.exp(-1/2), ((0.7)**2)*np.exp(-1/3), ((0.5)**2)*np.exp(-1/4), ((0.4)**2)*np.exp(-1/5), 0]
+    results = km.rate_for_nth_order(extent, T, A, Ea, n)
+    assert np.allclose(results, expected_results), "Rate for nth order reaction calculation failed."
+
+def test_rate_for_autocatalytic():
+    extent = np.array([0, 0.1, 0.3, 0.5, 0.6, 1])
+    T = np.array([100.0, 200.0, 300.0, 400.0, 500.0, 600.0])
+    A = 1.0  # Pre-exponential factor
+    Ea = 831.446261815324  # Activation energy in J/mol
+    n = 2  # Order of the reaction
+    m = 0.3 # Autocatalytic order
+    
+    expected_results = [0, (0.1**0.3)*((0.9)**2)*np.exp(-1/2), (0.3**0.3)*((0.7)**2)*np.exp(-1/3), (0.5**0.3)*((0.5)**2)*np.exp(-1/4), (0.6**0.3)*((0.4)**2)*np.exp(-1/5), 0]
+    results = km.rate_for_autocatalytic(extent, T, A, Ea, m, n)
+    assert np.allclose(results, expected_results), "Rate for nth order reaction calculation failed."
+
+def test_rate_for_kamal():
+    extent = np.array([0, 0.1, 0.3, 0.5, 0.6, 1])
+    T = np.array([100.0, 200.0, 300.0, 400.0, 500.0, 600.0])
+    A1 = 1.0  # Pre-exponential factor
+    E1 = 831.446261815324  # Activation energy in J/mol
+    A2 = 2  # Pre-exponential factor for second reaction
+    E2 = 415.723130907662  # Activation energy for second reaction in J/mol
+    n = 2  # Order of the reaction
+    m = 0.3 # Autocatalytic order
+    
+    expected_results = [np.exp(-1), 
+                        (np.exp(-1/2)+2*np.exp(-1/4)*0.1**0.3)*(0.9**2), 
+                        (np.exp(-1/3)+2*np.exp(-1/6)*0.3**0.3)*(0.7**2), 
+                        (np.exp(-1/4)+2*np.exp(-1/8)*0.5**0.3)*(0.5**2),
+                        (np.exp(-1/5)+2*np.exp(-1/10)*0.6**0.3)*(0.4**2),
+                        0]
+    results = km.rate_for_kamal(extent,T,A1,E1,A2,E2,m,n)
+    assert np.allclose(results, expected_results), "Rate for kamal reaction calculation failed."
+
+    
+def test_vitrification_WLF_rate():
[…]
