### 2025-06-26 0de9552
Slight modification of the GUI layout
