+++ kinopt/src/kinopt_interface.py
@@ -285,7 +285,7 @@ class Ui_MainWindow(object):
         MainWindow.setStatusBar(self.statusbar)
         self.dockWidgetData_extraction = QDockWidget(MainWindow)
         self.dockWidgetData_extraction.setObjectName(u"dockWidgetData_extraction")
-        self.dockWidgetData_extraction.setMinimumSize(QSize(352, 281))
+        self.dockWidgetData_extraction.setMinimumSize(QSize(356, 281))
         self.dockWidgetData_extraction.setFeatures(QDockWidget.AllDockWidgetFeatures)
         self.dockWidgetOpenFileContents = QWidget()
         self.dockWidgetOpenFileContents.setObjectName(u"dockWidgetOpenFileContents")
@@ -477,7 +477,7 @@ class Ui_MainWindow(object):
         self.toolBox_models_and_optimization_parameters.setFrameShape(QFrame.NoFrame)
         self.page_models_and_parameters = QWidget()
         self.page_models_and_parameters.setObjectName(u"page_models_and_parameters")
-        self.page_models_and_parameters.setGeometry(QRect(0, 0, 271, 747))
+        self.page_models_and_parameters.setGeometry(QRect(0, 0, 214, 747))
         self.verticalLayout_2 = QVBoxLayout(self.page_models_and_parameters)
         self.verticalLayout_2.setObjectName(u"verticalLayout_2")
         self.scrollArea_models_and_parameters = QScrollArea(self.page_models_and_parameters)
@@ -487,7 +487,7 @@ class Ui_MainWindow(object):
         self.scrollAreaWidgetContents_models_and_parameters = QWidget()
         self.scrollAreaWidgetContents_models_and_parameters.setObjectName(u"scrollAreaWidgetContents_models_and_parameters")
         self.scrollAreaWidgetContents_models_and_parameters.setEnabled(True)
-        self.scrollAreaWidgetContents_models_and_parameters.setGeometry(QRect(0, 0, 251, 680))
+        self.scrollAreaWidgetContents_models_and_parameters.setGeometry(QRect(0, 0, 194, 680))
         self.verticalLayout_3 = QVBoxLayout(self.scrollAreaWidgetContents_models_and_parameters)
         self.verticalLayout_3.setObjectName(u"verticalLayout_3")
         self.label_rate_model = QLabel(self.scrollAreaWidgetContents_models_and_parameters)
@@ -578,7 +578,7 @@ class Ui_MainWindow(object):
         self.toolBox_models_and_optimization_parameters.addItem(self.page_models_and_parameters, u"Models and parameters")
         self.page_optimization_methods_and_parameters = QWidget()
         self.page_optimization_methods_and_parameters.setObjectName(u"page_optimization_methods_and_parameters")
-        self.page_optimization_methods_and_parameters.setGeometry(QRect(0, 0, 271, 747))
+        self.page_optimization_methods_and_parameters.setGeometry(QRect(0, 0, 214, 747))
         self.verticalLayout_5 = QVBoxLayout(self.page_optimization_methods_and_parameters)
         self.verticalLayout_5.setObjectName(u"verticalLayout_5")
         self.scrollArea_optimization_methods = QScrollArea(self.page_optimization_methods_and_parameters)
@@ -588,7 +588,7 @@ class Ui_MainWindow(object):
         self.scrollArea_optimization_methods.setWidgetResizable(True)
         self.scrollAreaWidgetContents_optimization_methods = QWidget()
         self.scrollAreaWidgetContents_optimization_methods.setObjectName(u"scrollAreaWidgetContents_optimization_methods")
[…]
