### 2025-06-26 5bcac85
Correction of the product coupling function
