### 2025-06-26 7315104
Creation of the function and button to cancel the optimization
