### 2025-06-26 449d930
adding tests for coupling functions
