### 2025-06-26 #7751 @alan-tabore
Everything should be ready for review.
