### 2025-06-26 alan-tabore/KinOpt#4 @alan-tabore
I added some tests and will continue adding test (especially for the isoconversional analysis.

I also added GitHub Actions to run the tests automatically. Thanks again for the advice.
