### 2025-06-26 alan-tabore/KinOpt#2 @jranalli
The tests do indeed pass, but some don't do anything. See for example `test_optimization.py`. Line 15 actually doesn't find anything (i.e. `rss_functions` is empty).
https://github.com/alan-tabore/KinOpt/blob/449d93034678a64b48de49f5dedf31edc86977a5/kinopt/tests/test_optimization.py#L15

I think the line you want might be something like:
`rss_functions = [func for func in dict(inspect.getmembers(opt, inspect.isfunction)).values() if func.__name__.startswith('rss')]`
