### 2025-07-02 #7751 @rkurchin
@jranalli, thanks for this context, I generally agree – unit tests for a GUI seem unnecessary; testing that the backend works should be sufficient.

@mbkumar, have you had a chance to take another look at this?
