### 2025-07-21 alan-tabore/KinOpt#7 @mbkumar
Merge pacakges in requirements.txt into pyproject.toml
This is a simple python package without any submodules. All the dependencies can be put into pyproject.toml. It will simplify things for the end user.

From openjournals/joss-reviews#7751
