### 2025-07-21 alan-tabore/KinOpt#8 @mbkumar
Inconvenient launch of the kinopt GUI
Instead of asking users to go to `kinopt/src` and then run `python main.py`, the package can install python executables, such that users can launch the GUI with a singe command such as `kinopt`. Please refer to https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#creating-executable-scripts for additional documentation.

For https://github.com/openjournals/joss-reviews/issues/7751
