### 2025-08-05 #7751 @alan-tabore
Hello @rkurchin, I'll try to solve the issues raised by next Tuesday.
