### 2025-10-21 #7751 @mbkumar
The project is not following the software best practices. If the project
wants to be judged only on its scientific merit, it should be submitted to
a scientific journal. In my opinion, a joss submission helps in improving
the software aspects of a project with extensive software reviews followed
by fixes. I believe the project is still not up to the mark. It's lacking
in documentation and installation. I can't approve the publication of the
project in its current state.

I also don't want to review it further. Please remove me as reviewer.

Bharat Medasani
