### 2025-12-18 #7751 @mbkumar
Hello Rachel,

Sorry, getting swamped with other things. Will try to wrap it up this
weekend.

Bharat Medasani

On Thu, Dec 18, 2025 at 10:10 AM Rachel Kurchin ***@***.***>
wrote:

> *rkurchin* left a comment (openjournals/joss-reviews#7751)
> <https://github.com/openjournals/joss-reviews/issues/7751#issuecomment-3671023957>
>
> Hi @mbkumar <https://github.com/mbkumar>, another friendly ping on this!
>
