### 2025-12-23 #7751 @rkurchin
@alan-tabore, please also do the following:

0. Check on those DOI issues above.
1. Issue a new tagged release with all updates made during review.
2. Deposit that version of the code on an archival site (e.g. Zenodo); make sure the title, author list, version tag, and license match that of your JOSS submission.

Once you've done both of these things, send the version tag and DOI here! You can also directly rerun `@editorialbot check references` yourself to confirm the DOI fixes worked by posting a comment on this issue with that as the first line. (Changes only to the paper do not require a new version tag or archival deposit/DOI)
