### 2026-01-17 #7751 @rkurchin
It's just strange because that literally is the DOI that's listed on the [journal page](https://onlinelibrary.wiley.com/doi/full/10.1002/1096-987X%2820010130%2922%3A2%3C178%3A%3AAID-JCC5%3E3.0.CO%3B2-%23?casa_token=A0GIuK-J0DUAAAAA%3AS--2vO0hQP3zWfrlxUPDI3kmwAel5ql6fM-fkhmYxAetB-e_-h_UR-d8i112GvRqNvJ1gQYGDZuhZ-gR)...and when I click the link there, it seems to work. AFAICT, it's just that that `23` at the end somehow gets converted to a `#` in the PDF render.

Anyway, as I said above, @alan-tabore, let's just proceed with finalizing this in any case!
