### 2026-01-17 #7751 @danielskatz
The DOI on the journal page has a "%" in it - https://doi.org/10.1002/1096-987X(20010130)22:2%3C178::AID-JCC5%3E3.0.CO;2-%23 - while the DOI reported by check references doesn't
