### 2026-01-17 #7751 @danielskatz
👋 @tarleb - do you have any thoughts on this?

It seems like our {pandoc etc.} processing is having an issue with the "%" ...
