### 2026-01-18 #7751 @tarleb
According to https://github.com/jgm/pandoc/issues/8595, this should be fixed in pandoc. Not sure what's going on there.

Percent signs in URLs take on a special meaning, as they allow to encode characters which otherwise couldn't included in the URL, for example because they have a special meaning in an URL. A hash character cannot be part of a path in an URL, so it's percent-encoded as `%23`. There seems to be an issue somewhere in our stack, where the conversion between DOI and URI doesn't handle percent-encoded characters correctly.

I'll need to do some digging to find the source of this.
