### 2026-01-18 #7751 @tarleb
After some looking around, it seems that editorialbot does not invoke Inara for any of this, and @xuanxu is the most knowledgeable person regarding that part of the code base.
