### 2026-01-19 #7751 @xuanxu
I've updated our resolver to accept the `#` character in the DOI string.
Using the plain unencoded [DOI value](https://onlinelibrary.wiley.com/action/showCitFormats?doi=10.1002%2F1096-987X%2820010130%2922%3A2%3C178%3A%3AAID-JCC5%3E3.0.CO%3B2-%23) (`10.1002/1096-987X(20010130)22:2%3C178::AID-JCC5%3E3.0.CO;2-#`) shoud work now
