### 2026-01-19 #7751 @alan-tabore
Hello @rkurchin ,

I'm still struggling with the DOI.
I've tested both `10.1002/1096-987X(20010130)22:2%3C178::AID-JCC5%3E3.0.CO;2-#` and `10.1002/1096-987X(20010130)22:23C178::AID-JCC53E3.0.CO;2-23`
I'll just remove the citation if its okay with you.

I've released under the tag `v1.1` https://github.com/alan-tabore/KinOpt/tree/v1.1
and the doi is : [10.5281/zenodo.18305881](https://doi.org/10.5281/zenodo.18305881)
