### 2026-01-20 #7751 @arfon
@alan-tabore – you can also likely remove the DOI.

These DOIs are (in my opinion) a monstrosity with all the crazy characters 😸 , and while in the long run it would be good if we can fix this, I don't want to block this submission trying to debug this.
