### 2026-01-26 #7751 @rkurchin
Great, switching over to my AEiC hat to begin finalizing this...
As you can see, most seems in order, however the below are some points that require your attention (mostly copyediting things that shouldn't be too hard to address) :point_down: :
- [x] please ensure the Zenodo archive license matches your software license
- [x] line 11: add a comma after "extent of reaction"
- [x] line 12: "In most of the cases" -> "In most cases" and add a comma after the Arrhenius expression
- [x] 13: add a comma after "activation energy"
- [x] 14: remove comma after "But"
- [x] 14: "dependance" -> "dependence"
- [x] Suggest moving first paragraph of Statement of need to end of Summary, which otherwise does not currently mention the actual software. Also, remove the period after "optimization" and decapitalize "with" OR rewrite this second portion to actually be a full sentence (e.g. "It has a suite of features...")
- [x] Also, you never actually define isoconversional analysis, I would suggest adding a sentence (likely starting on current line 15) to briefly define it, or at least a reference
- [x] 31: suggest some sort of qualifier here; there are optimization tasks that are extremely computationally expensive, so presumably you mean that the optimizations being undertaken here do not need exceptional performance – also, this sentence is just overall odd as it sounds as if you selected Python on the basis of a negative quality, so perhaps a rephrase to something like "The Python programming language was selected due to its ubiquity and user-friendliness and on the basis that the optimizations required for [XXX your phrase here] do not necessitate exceptional computational performance."
- [x] 42: should there be a citation for Advanced Vyazovkin?
