### 2026-02-11 #7751 @alan-tabore
> Great, switching over to my AEiC hat to begin finalizing this...
> ## Checks on repository
>
>     * [x]  Project has OSI approved license[x]  Project features contributing guidelines
>
>
> ## Checks on review issue
>
>     * [x]  Review completed[x]  Software version tag listed here matches a tagged release
>
>
> ## Checks on archive
>
>     * [x]  Archive listed title and authors matches paper[x]  Archive listed license matches software license[x]  Archive listed version tag matches tagged release (and includes a potential `v`).
>
>
> ## Checks on paper
>
>     * [ ]  Checked paper formatting - including no text styling in title[x]  Check affiliations to make sure country acronyms are not used[x]  Checked reference rendering[x]  Checked if pre-print citations can be updated by published versions[ ]  Checked for typos
>
>
> ## Remaining points:
>
> As you can see, most seems in order, however the below are some points that require your attention (mostly copyediting things that shouldn't be too hard to address) 👇 :
>
>     * [ ]  please ensure the Zenodo archive license matches your software license[ ]  line 11: add a comma after "extent of reaction"[ ]  line 12: "In most of the cases" -> "In most cases" and add a comma after the Arrhenius expression[ ]  13: add a comma after "activation energy"[ ]  14: remove comma after "But"[ ]  14: "dependance" -> "dependence"[ ]  Suggest moving first paragraph of Statement of need to end of Summary, which otherwise does not currently mention the actual software. Also, remove the period after "optimization" and decapitalize "with" OR rewrite this second portion to actually be a full sentence (e.g. "It has a suite of features...")[ ]  Also, you never actually define isoconversional analysis, I would suggest adding a sentence (likely starting on current line 15) to briefly define it, or at least a reference[ ]  31: suggest some sort of qualifier here; there are optimization tasks that are extremely computationally expensive, so presumably you mean that the optimizations being undertaken here do not need exceptional performance – also, this sentence is just overall odd as it sounds as if you selected Python on the basis of a negative quality, so perhaps a rephrase to something like "The Python programming language was selected due to its ubiquity and user-friendliness and on the basis that the optimizations required for [XXX your phrase here] do not necessitate exceptional computational performance."[ ]  42: should there be a citation for Advanced Vyazovkin?

Hello @rkurchin , I should have everything sorted out.

License version on Zenodo = 1.1 and Software version on GitHub = 1.1

 * [x]  please ensure the Zenodo archive license matches your software license

 I changed the license on Zenodo to a BSD-3 license, as for the software.

* [x]  line 11: add a comma after "extent of reaction"
* [x]  line 12: "In most of the cases" -> "In most cases" and add a comma after the Arrhenius expression
* [x]  13: add a comma after "activation energy"
* [x]  14: remove comma after "But"
* [x]  14: "dependance" -> "dependence"
* [x]  Suggest moving first paragraph of Statement of need to end of Summary, which otherwise does not currently mention the actual software. Also, remove the period after "optimization" and decapitalize "with" OR rewrite this second portion to actually be a full sentence (e.g. "It has a suite of features...")
* [x]  Also, you never actually define isoconversional analysis, I would suggest adding a sentence (likely starting on current line 15) to briefly define it, or at least a reference

Definition of isoconversional analysis added on previous line 15) _This isoconversional analysis consists of retrieving the reaction parameters (temperature and reaction rate) at a specific stage of the chemical reaction in order to obtain the parameters of certain kinetic models (such as the activation energy of the reaction)._

* [x]  31: suggest some sort of qualifier here; there are optimization tasks that are extremely computationally expensive, so presumably you mean that the optimizations being undertaken here do not need exceptional performance – also, this sentence is just overall odd as it sounds as if you selected Python on the basis of a negative quality, so perhaps a rephrase to something like "The Python programming language was selected due to its ubiquity and user-friendliness and on the basis that the optimizations required for [XXX your phrase here] do not necessitate exceptional computational performance."

* [x]  42: should there be a citation for Advanced Vyazovkin?

This method corresponded to the citation that was not recognized during the citation check, so I deleted it. I just added a new one.
