### 2026-02-18 #7751 @rkurchin
Congrats @alan-tabore and thanks to reviewers @mbkumar and @jranalli, JOSS could not function without your efforts!
